feat: Extract navigation component and style into separate files
The commit refactors navigation related code by moving it into its own component (`Navigation.tsx`) and CSS module (`navigation.module.css`). This improves code readability and maintainability by separating concerns. Moreover, it also removes redundant code and enhances the clarity of the primary `header.tsx` and `header.module.css` files.
