+++ src/components/ui/header/navigation.module.css
@@ -1,0 +1,53 @@
+.navigation {
+  flex: 1 1 709px;
+
+  & ul {
+    container-type: inline-size;
+    display: flex;
+    height: 100%;
+
+    & li {
+      border-inline-end: var(--border);
+
+      @container (width < 709px) {
+        flex-grow: 1;
+      }
+
+      &:last-child {
+        margin-inline-start: auto;
+        border-inline-start: var(--border);
+        border-inline-end: none;
+      }
+
+      & a,
+      & span {
+        --gutter-x: 32px;
+
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        height: 100%;
+        padding-right: var(--gutter-x);
+        padding-left: var(--gutter-x);
+
+        @container (width < 709px) {
+          --gutter-x: 0;
+        }
+      }
+
+      & a {
+        &:focus-visible,
+        &:hover {
+          color: var(--color-secondary-4);
+          box-shadow: inset 0 -3px 0 var(--color-accent-1);
+        }
+      }
+
+      & span {
+        color: var(--color-secondary-4);
+        pointer-events: none;
+        box-shadow: inset 0 -3px 0 var(--color-accent-1);
+      }
+    }
+  }
+}
